fix(all): new warning flags and fixing errors

+++ src/array/sort_algorithms/bubble_sort.c
@@ -1,4 +1,5 @@
 #include <stddef.h>
+#include "dsa_array.h"
 #include "dsa_error.h"
 #include "dsa_extra.h"
 #include "../array.h"
@@ -12,12 +13,12 @@ ds_error_t ds_array_bubble_sort(void *array, unsigned length, size_t size,
 		return DS_INVALID_SIZE;
 
 	ds_byte_t *arr = array;
-	for (int i = 0; i < length - 1; i++) {
+	for (int i = 0; i < (int)length - 1; i++) {
 		ds_bool_t swapped = DS_FALSE;
-		for (int j = 0; j < length - 1 - i; j++) {
+		for (int j = 0; j < (int)length - 1 - i; j++) {
 			if (compare(arr + j * size, arr + (j + 1) * size) > 0) {
 				_ds_swap(arr + j * size, arr + (j + 1) * size,
-				      size);
+					 size);
 				swapped = DS_TRUE;
 			}
 		}
